requirements.txt
confluent-kafka==2.1.1
cassandra-driver==3.28.0
pandas==2.0.3
pyyaml==6.0.1
fastavro==1.8.4
requests==2.31.0
tqdm==4.66.1